pyproject.toml
[project]
name = "scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf>=1.7.2",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "reportlab>=4.4.1",
    "scikit-image>=0.25.2",
    "streamlit>=1.45.1",
]
